Period Aggregator

Create function that aggregates information over time periods, normalizes a count variable, and creates new fields to summarize this information.

Usage

PeriodAggregator(
  time_variable = "period_mid_time",
  period_width_variable = "num_days",
  count_variable = "cases_this_period",
  norm_variable = "population_reporting",
  rate_variable = "daily_rate",
  norm_exponent = 5
)

Arguments

time_variable

Name of the variable to characterize the temporal location of the time period.

period_width_variable

Name of variable to characterize the width of the time period.

count_variable

Name of variable to characterize the count variable being normalized.

norm_variable

Name of variable to be used to normalize the count variable.

rate_variable

Name of variable to be used to store the normalized count variable.

norm_exponent

Exponent to use in normalization. The default is `5`, which means `per 100,000`.