Wavelet Normalizer

Creates normalizing fields in `data`

Usage

WaveletNormalizer(
  time_variable = "period_end_date",
  series_variable = "deaths",
  trend_variable = "deaths",
  series_suffix = "_series",
  trend_suffix = "_trend",
  output_emd_trend = "emd_trend",
  output_norm = "norm",
  output_sqrt_norm = "sqrt_norm",
  output_log_norm = "log_norm",
  output_emd_norm = "emd_norm",
  output_emd_sqrt = "emd_sqrt",
  output_emd_log = "emd_log",
  output_detrend_norm = "detrend_norm",
  output_detrend_sqrt = "detrend_sqrt",
  output_detrend_log = "detrend_log",
  eps = 0.01
)

Arguments

time_variable

column name of time variable in `data`, default is "period_end_date"

series_variable

column name of series variable in `data`, default is "deaths_series"

trend_variable

column name of series variable in `data`, default is "deaths_trend"

series_suffix

suffix to be appended to series data fields

trend_suffix

suffix to be appended to trend data fields

output_emd_trend

name of output field for the empirical mode decomposition applied to `trend_variable`

output_norm

name of output field for the `series_variable` normalized by `output_emd_trend`

output_sqrt_norm

name of output field for the square root of `output_norm`

output_log_norm

name of output field for the logarithm of (`output_norm` + `eps`)

output_emd_norm

name of output field for the empirical mode decomposition applied to `output_norm`

output_emd_sqrt

name of output field for the empirical mode decomposition applied to `output_sqrt_norm`

output_emd_log

name of output field for the empirical mode decomposition applied to `output_log_norm`

output_detrend_norm

name of output field for the computed field `output_norm`-`output_emd_norm`

output_detrend_sqrt

name of output field for the computed field `output_sqrt_norm`-`output_emd_sqrt`

output_detrend_log

name of output field for the computed field `output_log_norm`-`output_emd_log`

eps

numeric value for normalized data to be perturbed by before computing the logarithm

Value

function that creates normalized trend and de-trended fields

## Returned Function

- Arguments * `data` data frame containing time series data - Return - `data` with additional normalized fields