Join user-defined lookup table

Joins user-defined lookup table to data

Usage

join_user_table(raw_data, user_table_path, lookup_type, join_by)

Arguments

raw_data

data frame of table to be harmonized

user_table_path

string indicating path to user-defined lookup table

lookup_type

string indicating type of lookup table (disease, location, sex). Used to determine columns to join by if join_by not specified

join_by

vector of strings indicating columns to join by (optional if lookup_type is disease, location, or sex)

Value

data frame of harmonized data with user-defined keys